pyproject.toml
[project]
name = "oe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gymnasium>=1.2.3",
    "openenv-core>=0.2.2",
]
